Segment code and add more comments <PD-763>
