+++ tasks.py
@@ -875,3 +875,20 @@ def process_sendgrid_event(events):
     event_list = get_event_list(events)
     events_to_create = event_list_to_email_log(event_list)
     EmailLog.objects.bulk_create(events_to_create)
+
+
+@task(name="tasks.send_event_email", ignore_result=True)
+def send_event_email(email_task):
+    """
+    Send an appropriate email given an EmailTask instance.
+
+    Inputs:
+    :email_task: EmailTask we are using to generate this email
+    """
+    email_task.task_id = send_event_email.request.id
+    email_task.save()
+
+    email_task.send_email()
+
+    email_task.completed_on = datetime.now()
+    email_task.save()
